+++ StepTrackerTCA/Features/Records/AddMeasurement/Feature/AddMeasurementFeature.swift
@@ -62,105 +62,3 @@ struct AddMeasurementFeature {
     }
     
 }
-
-import ComposableArchitecture
-import Foundation
-
-/// Implementation of `AddMeasurementFeature` action
-extension AddMeasurementFeature {
-    
-    @CasePathable
-    enum Action: ViewAction, BindableAction {
-        
-        // MARK: - Binding Action
-        
-        /// Handles changes in bindings for the state.
-        case binding(BindingAction<State>)
-        
-        // MARK: - Actions
-        
-        /// Triggered when the user selects a different workout type in the movement picker.
-        case selectedWorkoutPickerChange(WorkoutType?)
-      
-        ///
-        case selectedWeightliftingMovementPickerChange(WeightliftingMovement?)
-        
-        ///
-        case selectedStrengthMovementPickerChange(StrengthMovement?)
-        
-        ///
-        case selectedFitnessMovementPickerChange(FitnessMovement?)
-        
-        ///
-        case selectedCrossMovementPickerChange(CrossMovement?)
-        
-        ///
-        case selectedHeroMovementPickerChange(HeroMovement?)
-        
-        ///
-        case selectedWeightUnitPickerChange(WeightUnit)
-        
-        ///
-        case selectedWorkoutUnitPickerChange(WorkoutUnit)
-        
-        // MARK: - View Actions
-        
-        case view(View)
-        
-        enum View {
-            
-            case viewDidAppear
-        }
-    }
-    
-}
-
-
-import ComposableArchitecture
-import Foundation
-
-/// Implementation of `AddMeasurementFeature` state
-extension AddMeasurementFeature {
-    
-    @ObservableState
-    struct State: Equatable {
-        
-        // MARK: - Properties
-        
-        /// The date for which the user wants to add  data.
-        /// Defaults to the current date.
-        var addDataDate: Date = .now
-        
-        ///
-        var selectedTime: Date = .now
-        
-        ///
-        var workoutType: WorkoutType? = nil
-        
-        ///
-        var weightliftingMovement: WeightliftingMovement? = nil
-        
-        ///
-        var strengthMovement: StrengthMovement? = nil
-        
-        ///
-        var fitnessMovement: FitnessMovement? = nil
-        
-        ///
-        var crossMovement: CrossMovement? = nil
-        
-        ///
-        var heroMovement: HeroMovement? = nil
-        
-        /// The value to add, entered as a string.
-        var valueToAdd: String = ""
-        
-        /// The selected unit of measurement for weight.
-        var weightUnit: WeightUnit = .kg
-        
-        ///
-        var workoutUnit: WorkoutUnit = .reps
-        
-    }
-    
-}
